Indian-Sign-Language-Gesture-Recognition-master/ISL-Recognition-Modern/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
greenlet==3.0.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Machine Learning & Computer Vision
tensorflow==2.15.0
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.3
scikit-learn==1.4.0
pillow==10.2.0

# Audio Processing
SpeechRecognition==3.10.1
pyttsx3==2.90
pydub==0.25.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
httpx==0.26.0
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
